requirements.txt
asgiref==3.2.7
Django==3.0.5
django-braces==1.14.0
djangorestframework==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
